+++ lib/metro/stubs/page_stub.dart
@@ -3,7 +3,7 @@ import 'package:recase/recase.dart';
 String pageStub({required ReCase pageName}) => '''
 import 'package:flutter/material.dart';
 import 'package:nylo_framework/nylo_framework.dart';
-import '../../app/controllers/controller.dart';
+import '/app/controllers/controller.dart';
 
 class ${pageName.pascalCase}Page extends NyStatefulWidget {
   final Controller controller = Controller();
@@ -17,6 +17,7 @@ class _${pageName.pascalCase}PageState extends NyState<${pageName.pascalCase}Pag
 
   @override
   init() async {
+    super.init();
     
   }