+++ lib/metro/stubs/page_w_controller_stub.dart
@@ -1,11 +1,9 @@
 import 'package:recase/recase.dart';
 
-String pageWithControllerStub(
-        {required ReCase className, required String importName}) =>
-    '''
+String pageWithControllerStub({required ReCase className}) => '''
 import 'package:flutter/material.dart';
 import 'package:nylo_framework/nylo_framework.dart';
-import '../../app/controllers/${importName.toLowerCase()}_controller.dart';
+import '/app/controllers/${className.snakeCase}_controller.dart';
 
 class ${className.pascalCase}Page extends NyStatefulWidget {
   final ${className.pascalCase}Controller controller = ${className.pascalCase}Controller();
@@ -20,7 +18,7 @@ class _${className.pascalCase}PageState extends NyState<${className.pascalCase}P
 
   @override
   init() async {
-  
+    super.init();
   }
   
   @override