+++ lib/metro/stubs/postman_api_service_stub.dart
@@ -1,7 +1,10 @@
 import 'package:recase/recase.dart';
 
 String postmanApiServiceStub(ReCase rc,
-    {required String networkMethods, required, required String imports, required String baseUrl}) =>
+        {required String networkMethods,
+        required,
+        required String imports,
+        required String baseUrl}) =>
     '''
 ${baseUrl == "getEnv('API_BASE_URL')" ? "import 'package:nylo_framework/nylo_framework.dart';\n" : ""}import 'package:flutter/material.dart';
 ${imports != "" ? imports + '\n' : ''}import '/app/networking/dio/base_api_service.dart';